+++ src/compute-engine/index.ts
@@ -175,6 +175,7 @@ export {
   simplify,
   evaluate,
   N,
+  declare,
   assign,
   expand,
   expandAll,